Add documentation and ImageUtil crop function to attempt remove black bars in thumbnail views.
