+++ iOS-Meme-Version2/Controllers/SentMemeCollectionViewCell.swift
@@ -8,6 +8,8 @@
 import Foundation
 import UIKit
 
+// MARK : Represents the image view cell of a meme.
+
 class SentMemeCollectionViewCell: UICollectionViewCell {
     
     @IBOutlet weak var sentMemeImageView: UIImageView!
